#39 Fixed typo

+++ proj/src/main/java/com/essentls/dao/UserPaymentsListDAO.java
@@ -49,7 +49,7 @@ public final class UserPaymentsListDAO extends AbstractDAO<List<Payment>> {
 
         try {
             stmt = con.prepareStatement(STATEMENT);
-            stmt.setInt(1, this.user.getId());
+            stmt.setLong(1, this.user.getId());
 
             rs = stmt.executeQuery();
